+++ src/locale/NodeTexts.ts
@@ -205,7 +205,7 @@ type NodeTexts = {
             /** When there's no ending expression */
             ExpectedEndingExpression: InternalConflictText;
             /** When a statement is ignored because it's not last and not a bind */
-            IgnoredExpression: ConflictText;
+            IgnoredExpression: ConflictText & { resolution: Template };
         }>;
     /**
      * A single boolean literal, e.g., `⊤` or `⊥`